Add explicit expectations between employees and the org
We already list our values in the Handbook but I think it’s important for us to be explicit about the healthy tension between the expectations of employees and the organisation.

These expectations also set the context for escalation procedures for employees to raise an issue and also for the organisation to escalate issues with employees.

**Further Background**

One of our People objectives for Q2 is to provide an escalation procedure in the Handbook when expectations aren’t being met. We already have this for employees to raise an issue, we are now looking at it from the other way when an employee isn’t meeting the expectations of their role.

A future PR will include the expectation health check linked to along with more formal capability procedures.