Add guidance around only one person away per team at a time
We’ve had several issues with customers that haven’t been helped by multiple people being away on a team at a time. We add a guideline that only one person should be away at a time on a particular team. Anyone wanting to book holiday outside of this guideline will need to escalate the request to a director.