Create Internal delivery team support process
What:
This is a guide to requesting support of delivery team members for sales opportunities or internal projects.

Why we want to introduce this process:
Spinning up internal delivery teams is a difficult process that requires clarity over the details before it can take place. These processes aim to enable team members requiring support to provide the necessary information to the Operations Team to get their requests fulfilled as efficiently and effectively as possible. It also ensures that the details of your request go to a consistent location and is not lost in slack direct messages.