Creating and changing delivery teams process
**What:**
This is a guide to submitting a request to the Operations Team about starting a new delivery team for a new customer, or changing the structure of an existing delivery team.

**Why we want to introduce this process:**
Spinning up a delivery team support is a process that requires clarity over the details before it can take place. These processes aim to enable the Operations Team to get new delivery teams started as efficiently and effectively as possible. It also ensures that the details of your request go to a consistent location and are not lost in slack direct messages.