Added information about dishwasher use and rubbish
I've written up some guidelines around keeping tableware clean, and
using the dishwasher. I think this is important as the company grows and
I've noticed piles of dirty tableware and lack of glasses lately.

I've also added some information on using the bins. Sometimes they get
really full and no one is too good to take out the rubbish!

Please let me know if any of this needs amending because the new
cleaners take care of it. I'm not sure what the arrengement with them
is.