Create device_admin_profiles.md
As part of our work towards Cyber Essentials + and ISO27001 certification, we have been advised to implement the separation of admin and standard user profiles on Made Tech devices, specifically those being used by business teams. We agree with the rationale behind this implementation, but are keen to see it does not block team members from installing software necessary to the performance of their job, and does not generate additional work for the ISMS team in being responsible for manually granting admin permissions for software installation. 

We have tried to develop this policy with a heavy emphasis on trusting the team to act responsibly when downloading and installing software. If you feel this policy restricts your ability to work effectively, please get in contact with either the CSO (Chris) or the ISMS Manager (Emily).