Delete code definitions of roles and expectations
## Why

We have and are moving to using the SFIA framework which supersedes defining expectations here, and removes the need for JobSpec (code library for specifying role expectations).

We also have problems maintaining these roles because they require knowledge of code to edit which has blocked our people and operations teams from working with them so easily.

## What

We remove all the code behind the MD files. You must for now edit the MD files themselves.