Update learning pages to be inclusive and explain Learn Tech
Following feedback in issue #318 the handbook suggested that Learn Tech
was for Engineering staff only, which is completely wrong.

This led to realising that the handbook also referenced old Learn Tech
and even including a spreadsheet that included some fairly old data.

Even though engineers are important (no more important than anyone else in
the company though!), the specific learn tech and onboarding rather than
being worthy of being the first thing on the page, and the mentioned
activities are more of an exception to regular staff activities, I moved
them further down the order of information.