Add SFIA expectations for technology roles
- Add role summary and required competencies for all technology roles
- Add SFIA examples to all technology roles
- Add links to SFIA guidance from roles/README.md
- Include navigation between each SFIA level
- Include links to official SFIA guidance for each level

This change sees us move SFIA guidance out of documents and spreadsheets and into the handbook where it will be more accessible.

Since introducing SFIA we've received feedback that it isn't clear how to use SFIA guidance for roles and that each role has so many examples it is hard to know what good looks like for a particular role.

In order to make technology SFIA guidance clearer for each role, we've provided single sentence (in most cases) summaries of autonomy, influence, complexity, nowledge and business skills within our organisation and public sector industry context.

We've kept examples from existing SFIA guidance but have put these at the end of each summary and hopefully clearly convey that these are provided as illustrative examples only and should not be used as checklists.

The people team are in the process of putting guidance together for line managers and those seeking to develop their careers on how SFIA can be used for probation and performance reviews, pay rises and promotions.