Guidance for 121s
A first version of basic 121 guidance. This speaks to Line Managers and the people who they line manage. Here called 'Direct reports' for lack of a better word.

The aim is to keep this guidance as brief as possible. 
There are of course more things we use 121s for, such as building a good relationship and get to know each other and check in on peoples well-being. 
I think this page could quickly become a long read so wanted to just get the basics written up here. Thoughts?