Add explanation for specialist technology roles
# What

Add explanation as to why you may see cloud or data engineer roles advertised but won't find them in the Handbook.

# Why

We advertise specialist cloud and data engineering roles but this does not reflect career pathways at Made Tech. Internally we have a single career pathway for engineering where individuals can build out a breadth or depth or a mixture of both expertise. The projects individuals work on will be based on experience and training rather than job titles.