Add explanation for specialist technology roles (#504)
* Add explanation for specialist technology roles

# What

Add explanation as to why you may see cloud or data engineer roles advertised but won't find them in the Handbook.

# Why

We advertise specialist cloud and data engineering roles but this does not reflect career pathways at Made Tech. Internally we have a single career pathway for engineering where individuals can build out a breadth or depth or a mixture of both expertise. The projects individuals work on will be based on experience and training rather than job titles.

* Remove cloud engineer role from Handbook

# What

We remove the cloud engineer role.

# Why 

We do not have specialist career pathways, only a single software engineer pathway that provide our engineers the freedom to explore a variety of technologies and specialisms.