Changing Flexible Leave to Fixed Leave
Flexible leave will end on 31/12/200
This update is to update the guidance to align with our updated holiday benefit which will make a fixed pot of leave available to all team members. The total holiday available is set to 30 days between June and May, changing from January to December, which is to align it to our financial year.