Update how_to_timesheet.md (#904)
Minor amendment to clarify that 35 hours of timesheet time expected each week reflect billable time. 

Further work needed in future to update this page to reflect updates to time categories; and create greater clarity on the relationship between 'billable hours' and internal activities on top of these. Ops team to pick this up in the next 1-2 months.